Put store inventories in a separate module
